Merge build and install jobs
Ideally, the two jobs should remain separated. The `install` job would
use almost the same `matrix` (minus `target`) as the `build` job and
have the same `needs` and `if` as the `test` job. This would however
make a duplicated and error-prone `matrix` for the two jobs, and sharing
a `matrix` between jobs is currently not possible with GitHub Actions
(neither directly, nor using YAML anchors, other options were too
 cumbersome when writing this). Consequently, for the sake of
maintainability, the two jobs are merged for now.